Workers module
Separated thread Workers into separate module
